LPS-115381 Moving to package info.item.updater
