+++ modules/apps/translation/translation-api/bnd.bnd
@@ -4,4 +4,4 @@ Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.translation.exception,\
 	com.liferay.translation.exporter,\
-	com.liferay.translation.info
+	com.liferay.translation.info.item.updater
