+++ modules/apps/translation/translation-api/src/main/resources/com/liferay/translation/info/item/updater/packageinfo
