+++ modules/apps/translation/translation-test/src/testIntegration/java/com/liferay/translation/info/item/updater/test/InfoFormValuesUpdaterArticleTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.translation.exporter.test;
+package com.liferay.translation.info.item.updater.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.info.field.InfoFormValues;
@@ -37,7 +37,7 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.translation.exporter.TranslationInfoFormValuesExporter;
-import com.liferay.translation.info.InfoFormValuesUpdater;
+import com.liferay.translation.info.item.updater.InfoFormValuesUpdater;
 import com.liferay.translation.test.util.TranslationTestUtil;
 
 import java.util.Locale;
